requirements.txt
langchain==0.0.284
python-dotenv==1.0.0
streamlit==1.22.0
youtube-transcript-api
openai
protobuf~=3.19.0